+++ packages/terra-form-select/src/MultiSelectField.jsx
@@ -41,11 +41,15 @@ const propTypes = {
    */
   intl: intlShape.isRequired,
   /**
+   * Whether the field displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
+   */
+  isIncomplete: PropTypes.bool,
+  /**
    * Whether the field is displayed inline. Displays block by default.
    */
   isInline: PropTypes.bool,
   /**
-   * Whether the field is invalid.
+   * Whether the field displays as Invalid. Use when value does not meet validation pattern.
    */
   isInvalid: PropTypes.bool,
   /**
@@ -110,6 +114,7 @@ const defaultProps = {
   error: undefined,
   help: undefined,
   hideRequired: false,
+  isIncomplete: false,
   isInline: false,
   isInvalid: false,
   isLabelHidden: false,
@@ -133,6 +138,7 @@ const MultiSelectField = ({
   help,
   hideRequired,
   intl,
+  isIncomplete,
   isInline,
   isInvalid,
   isLabelHidden,
@@ -203,6 +209,7 @@ const MultiSelectField = ({
         aria-describedby={ariaDescriptionIds}
         disabled={selectAttrs.disabled || disabled}
         id={selectId}
+        isIncomplete={isIncomplete}
         isInvalid={isInvalid}
         isTouchAccessible={isTouchAccessible}
         defaultValue={defaultValue}