+++ packages/terra-form-select/src/terra-dev-site/test/form-select/RequiredDefault.test.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Select from '../../../Select';
+import styles from './common/Select.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class ControlledDefault extends React.Component {
   constructor() {
@@ -15,19 +19,21 @@ class ControlledDefault extends React.Component {
 
   render() {
     return (
-      <Select
-        placeholder="Select a color"
-        required
-        onChange={this.handleChange}
-        id="default"
-        value={this.state.value}
-      >
-        <Select.Option value="blue" display="Blue" />
-        <Select.Option value="green" display="Green" />
-        <Select.Option value="purple" display="Purple" />
-        <Select.Option value="red" display="Red" />
-        <Select.Option value="violet" display="Violet" />
-      </Select>
+      <div className={cx('content-wrapper')}>
+        <Select
+          placeholder="Select a color"
+          required
+          onChange={this.handleChange}
+          id="default"
+          value={this.state.value}
+        >
+          <Select.Option value="blue" display="Blue" />
+          <Select.Option value="green" display="Green" />
+          <Select.Option value="purple" display="Purple" />
+          <Select.Option value="red" display="Red" />
+          <Select.Option value="violet" display="Violet" />
+        </Select>
+      </div>
     );
   }
 }