+++ packages/terra-form-textarea/src/TextareaField.jsx
@@ -46,11 +46,15 @@ const propTypes = {
   // eslint-disable-next-line react/forbid-prop-types
   inputAttrs: PropTypes.object,
   /**
+   * Whether the field displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
+   */
+  isIncomplete: PropTypes.bool,
+  /**
    * Whether or not the field is an inline field.
    */
   isInline: PropTypes.bool,
   /**
-   * Whether or not the field is invalid.
+   * Whether the field displays as Invalid. Use when value does not meet validation pattern.
    */
   isInvalid: PropTypes.bool,
   /**
@@ -100,6 +104,7 @@ const defaultProps = {
   help: null,
   hideRequired: false,
   inputAttrs: {},
+  isIncomplete: false,
   isInline: false,
   isInvalid: false,
   isLabelHidden: false,
@@ -122,6 +127,7 @@ const TextareaField = (props) => {
     hideRequired,
     inputAttrs,
     inputId,
+    isIncomplete,
     isInline,
     isInvalid,
     isLabelHidden,
@@ -171,6 +177,7 @@ const TextareaField = (props) => {
         {...inputAttrs}
         disabled={inputAttrs.disabled || disabled}
         id={inputId}
+        isIncomplete={isIncomplete}
         onChange={onChange}
         placeholder={placeholder || inputAttrs.placeholder}
         value={value}