+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/AutoResizableTextareaMobileOnTouchStart.test.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Textarea from '../../../Textarea';
+import styles from './common/Textarea.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 export default class textarea extends React.Component {
   constructor() {
@@ -25,7 +29,9 @@ export default class textarea extends React.Component {
 
   render() {
     return (
-      <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" ariaLabel="label" />
+      <div className={cx('content-wrapper')}>
+        <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" ariaLabel="label" />
+      </div>
     );
   }
 }