+++ packages/terra-form-field/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Updated `isInvalid` prop description.
 
 3.26.0 - (October 3, 2019)
 ------------------